server/requirements.txt
django==2.1.2
djangorestframework==3.9.0
whitenoise==4.1
django-celery-results==1.0.4
redis==2.10.6
docker==3.5.1 
django_celery_beat==1.3.0
django-request-logging==0.6.6
numpy==1.15.4
